# swagger.yaml
openapi: 3.0.0
info:
  title: Node JWT Authentication API
  description: API for user registration, login, and dashboard access
  version: 1.0.0
paths:
  /:
    get:
      summary: Check if the server is up and running
      tags:
        - "Authentication"
      responses:
        '200':
          description: Server is up and running
  /register:
    post:
      summary: Register a new user
      tags:
        - "Authentication"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstname:
                  type: string
                  example: John
                lastname:
                  type: string
                  example: Doe
                email:
                  type: string
                  example: "guest@guest.com"
                password:
                  type: string
                  example: "guest123"
              required:
                - firstname
                - lastname
                - email
                - password
      responses:
        '201':
          description: User registered successfully
        '400':
          description: Bad request - All input is required or user already exists
        '500':
          description: Internal Server Error
  /login:
    post:
      summary: Log in an existing user
      tags:
        - "Authentication"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: "guest@guest.com"
                password:
                  type: string
                  example: "guest123"
              required:
                - email
                - password
      responses:
        '200':
          description: User logged in successfully
        '400':
          description: Bad request - All input is required or invalid credentials
        '500':
          description: Internal Server Error
  /dashboard:
    get:
      summary: Get user dashboard data
      security:
        - BearerAuth: []
      tags:
        - "Dashboard"
      responses:
        '200':
          description: User dashboard data retrieved successfully
        '404':
          description: User not found
        '500':
          description: Internal Server Error
  
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer